+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- (<= MC1.21) [back] fix: the `/back` command should intercept all types of teleport. (e.g. the `/tp` command.) (Thanks to @itzrobotix)
+- [world] refactor: now `/world tp` simply teleports to the `target dimension` with the same coordinate, instead of start a rtp. (The rtp is likely to failed, and cost too much.)